+++ src/Agents.jl
@@ -71,6 +71,8 @@ Noteworthy changes:
 - Agents.jl moved to Julia 1.9+, and now exports visualization
   and interactive applications automatically once Makie (or Makie backends
   such as GLMakie) come into scope, using the new package extension system.
+  The only downside of this is that now to visualize ABMs on open street
+  maps, the package OSMMakie.jl must be explicitly loaded as well.
 - Several performance improvements all across the board.
 - DEI-motivated name change for all names that remove agents:
     - `genocide! -> remove_all!`