+++ src/visualizations.jl
@@ -63,6 +63,7 @@ See also [`abmvideo`](@ref) and [`abmexploration`](@ref).
 * `graphplotkwargs = NamedTuple()` : keywords directly passed to
   [`GraphMakie.graphplot!`](https://graph.makie.org/stable/#GraphMakie.graphplot)
   if model space is `GraphSpace`.
+* `adjust_aspect = true`: Adjust axis aspect ratio to be the model's space aspect ratio.
 
 The stand-alone function `abmplot` also takes two optional `NamedTuple`s named `figure` and
 `axis` which can be used to change the automatically created `Figure` and `Axis` objects.